+++ CHANGELOG.md
@@ -43,6 +43,7 @@ TODO: Date
   * `with { …; } <…>` expressions to insert statements into the `.render` method
   * Conditional content via `if {…} <…>`, `if …… else <…>` and `match <…> [ … ]`
   * Box expressions: `box ⟦priv …⟦: ⟦struct⟧ … ⟦where …;⟧⟧⟧ <…>`
+  * More accurate `Unpin` constraints on generated storage context types
 
 * Revisions:
   * Improved `Counter` example in the README.